Refactor tests for post-processing
